requirements.txt
streamlit
streamlit-autorefresh
pandas
plotly
pyserial
paho-mqtt
firebase-admin
python-dotenv